backend/requirements.txt
Django == 4.0.5
djangorestframework == 3.13.1
psycopg2 == 2.9.3
django-filter == 23.2
django-cors-headers == 3.14.0
drf-yasg==1.21.7
pillow == 10.3.0
torch == 2.2.0
torchvision == 0.17.0
numpy == 1.26.4
pytesseract==0.3.10
opencv-python==4.7.0.72
scikit-learn==1.3.0
ultralytics==8.1.18